Git::Deploy::Timing: really fix exit values
I have no idea what's going on with this END {} block, but I still
have the issue of successful "start" and "abort"'s exiting with e.g. 1
because we get 256 in here for some reason.

Just make this a bit less robust but make it actually work wrt access
codes. This continues the hack I started in a8d0e1d.