Make the "revert" action take an argument
Now the revert action can take an argument "revert 2" is revert to the
last rollout, "revert quit" is a no-op etc. This uses the same
validation logic for the argument as the existing code.

The main purpose of this is to be able to test the revert function, we
now have tests to make sure it works.