Make it so we can do "dzil test" without errors from dzil
The tests do not actually pass when run this way currently, but
that is a separate issue. We can now use make test to test from within
the git-deploy repo, and use dzil test to test what it would be like
to run the tests from a distribution without .git in it.