Phase Hooks: introduce a new GIT_DEPLOY_PREFIX env variable
Pass the prefix we're using internally to hooks as GIT_DEPLOY_PREFIX,
this allows us to have hooks in "common" that will still know what
prefix they're dealing with, hooks can still use the
"GIT_DEPLOY_HOOK_PREFIX" to get either the real prefix or "common"
depending on where they are.

As it turns out we internally had a hook that was trying to use
"GIT_DEPLOY_PREFIX" which didn't work until this commit, I've read the
code and validated that it'll still work as intended.

Also, document all these environment variables that we're passing to
the hooks. We still need to document the sync hook variables, but we
can do that later.