Git::Deploy::Say: don't refuse to work if we can't log
Instead of dying if we can't log just warn once per invocation
instead, just because the permissions on our logfile are screwed that
doesn't mean we should halt all deployments.

In addition this logging code was overly verbose because it was trying
only to open the log handle and computing the log filename once per
process. We can just use Memoize for that and save ourselves the
verbosity.