Git::Deploy: don't implictly read user.* variables as --global
Change the user.* variables so we don't force them to be read
globally. It's not an illegitimate use case to set a username/email
per rollout repository, or maybe to have some strange setup where each
user deploys from their home directory.

The immediate need I have for this is to be able to test how the tool
behaves with this unset without editing ~/.gitconfig, but I don't see
the harm in just never adding --global here.