Actually support Dist::Zilla in some non-trivial capacity
Finish up the support for Dist::Zilla, now we have (in theory)
complete Dist::Zilla integration where you can do:

    dzil release

And we'll make a release, bump the changelog & commit it, create a
tag, and then push it to origin.

This in effect reverts commit 9269fe9bd5567f73aed1ebb3b5c698e861218733
of adding a Makefile.PL.It doesn't make any sense to check this file
in, with dist.ini this'll be created with "dzil build", and to test we
should be doing "dzil test" instead of "make test".