Push any tags we create right away to the remote
Writing a sync hook with the semantics of:

 1. You do git-deploy sync
 2. The sync hook reads your .deploy info, then syncs your tag out
    directly from upstream.

Currently isn't possible because we don't push the tag before we sync.

This also makes other things confusing, e.g. you don't see tags in any
commit stream to the upstream repo until the whole sync is complete,
so e.g. if the sync fails partway through you can't grab the failed
tag from the upstream to debug it locally.

The tests pass with this, but I haven't added new ones or tested this
in production, so I'm not pushing it to master for now. This is
tracked as Issue #24.