Untangle issues with END blocks, and exit codes (introduce fatal_exit)
When $? is set inside of an END block it changes the exit code returned
to the OS. (See the docs for $? in perlvar.)

We were calling routines that would execute system/backticks inside of
an END block in Git::Deploy::Timing, which would clobber the exit code.

Various patches ensued to try to work around this fact, but ultimately
confused the matter more. This patch effectively reverts those patches,
and makes it so we get configuration for Git::Deploy::Timing at startup
and not shutdown, thus leaving the exit code alone during teardown.

This patch also introduces a new sub exported by Git::Deploy::Say called
fatal_exit($code,@msg) which is intended for top level routines to use for
fatal exits which should return a specific exit code. Internal routines
should continue to use _die() as it is trapable wheras fatal_exit() is
not. In the long run we need to rething this separation of
functionality.