Add a "push-hotfix" action to automate hotfix recipe
See Issue #25 - TODO Eliminate busywork from hotfix process.

This is an untested initial attempt at a recipe for this process. It is
interesting as we dont have an easy way currently to the lock a directory
without modifying the logfile. We solve this by holding a lock on the
logfile for the duration of the action.

IOW, this includes code to support a "logless lock". (Quite simple logic
actually).